♻️ Refactor(folder) : Modal 컴포넌트 & 옵저버 영역 리팩토링
- 타입 별로 나뉘어져 있던 Modal 컴포넌트를 하나로 통일
- ModalPortal 여러번 import 되는 이슈 해결
- 링크 추가하기 영역 value 서로 연동되도록 state 추가
