--- components/common/Modal/ShareModal.tsx
@@ -1,71 +1,0 @@
-import styled from 'styled-components';
-import ModalFrame from './ModalFrame';
-import { SHARE_LIST } from '@/constants/default';
-import { shareKakaotalk, shareFacebook, copyLink } from '@/lib/utils/shareLink';
-import { MouseEvent } from 'react';
-import Image from 'next/image';
-
-interface Props {
-  data: string;
-  folderId: number;
-  onClickClose: () => void;
-}
-
-function ShareModal({ data = '', folderId, onClickClose }: Props) {
-  const SHARE_URL = `shared?user=1&folder=${folderId}`;
-
-  function handleShareClick(event: MouseEvent): void {
-    const el = event.target as HTMLElement;
-    const target = el?.closest('div')?.children[1];
-    if (target?.textContent === '카카오톡') shareKakaotalk(data, SHARE_URL);
-    if (target?.textContent === '페이스북') shareFacebook(SHARE_URL);
-    if (target?.textContent === '링크 복사') copyLink(SHARE_URL);
-  }
-
-  return (
-    <ModalFrame onClickClose={onClickClose}>
-      <Title>폴더 공유</Title>
-      <FolderName>{data}</FolderName>
-      <Box>
-        {SHARE_LIST.map((item) => (
-          <Item key={item.msg}>
-            <Icon src={item.src} alt={item.alt} onClick={handleShareClick} />
-            <Text>{item.msg}</Text>
-          </Item>
-        ))}
-      </Box>
-    </ModalFrame>
-  );
-}
-
-export default ShareModal;
-
-const Title = styled.div`
-  font-size: 2rem;
-  font-weight: 700;
-`;
-const FolderName = styled.div`
-  color: var(--gray-60);
-  font-size: 1.4rem;
-  line-height: 2.2rem;
-`;
-const Box = styled.div`
-  margin-top: 24px;
-  display: flex;
-  gap: 32px;
-`;
-const Item = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  align-items: center;
-`;
-const Icon = styled(Image)`
-  &:hover {
-    cursor: pointer;
-  }
-`;
-const Text = styled.div`
-  font-size: 1.3rem;
-  line-height: 1.5rem;
-`;
